Move more system test functionality into the testhelpers package
