+++ system_tests/gitlab/_helpers/asserts.py
@@ -5,8 +5,11 @@ import pathlib
 import gitlab
 import repobee_plug as plug
 
-from .const import ORG_NAME, LOCAL_BASE_URL, TOKEN, TEACHER, TASK_CONTENTS_SHAS
-from .helpers import hash_directory, get_group
+from repobee_testhelpers._internal.templates import TASK_CONTENTS_SHAS
+from repobee_testhelpers.funcs import hash_directory
+
+from .const import ORG_NAME, LOCAL_BASE_URL, TOKEN, TEACHER
+from .helpers import get_group
 
 
 def assert_template_repos_exist(assignment_names, org_name):
