+++ system_tests/gitlab/_helpers/helpers.py
@@ -1,9 +1,7 @@
 """Helper functions for the integration tests."""
-import hashlib
 import os
 import pathlib
 import subprocess
-import sys
 import tempfile
 
 from typing import Optional
@@ -47,32 +45,6 @@ def get_group(group_slug: str, gl: Optional[gitlab.Gitlab] = None):
         for group in gl.groups.list(search=group_slug)
         if group.path == group_slug
     ][0]
-
-
-def run_in_docker_with_coverage(command, extra_args=None):
-    assert extra_args, "extra volume args are required to run with coverage"
-    coverage_command = (
-        "coverage run --branch --append --source _repobee -m " + command
-    )
-    return run_in_docker(coverage_command, extra_args=extra_args)
-
-
-def run_in_docker(command, extra_args=None):
-    extra_args = " ".join(extra_args) if extra_args else ""
-    docker_command = (
-        "sudo docker run {} --net development --rm --name repobee "
-        "repobee:test /bin/sh -c '{}'"
-    ).format(extra_args, command)
-    proc = subprocess.run(
-        docker_command,
-        shell=True,
-        stdout=subprocess.PIPE,
-        stderr=subprocess.STDOUT,
-    )
-    print(
-        proc.stdout.decode(sys.getdefaultencoding())
-    )  # for test output on failure
-    return proc
 
 
 def update_repo(repo_name, filename, text):
@@ -120,18 +92,6 @@ def update_repo(repo_name, filename, text):
         assert push_proc.returncode == 0
 
     assert proj.files.get(filename, "master").decode().decode("utf8") == text
-
-
-def hash_directory(path):
-    shas = []
-    for dirpath, _, filenames in os.walk(str(path)):
-        if ".git" in dirpath:
-            continue
-        files = list(
-            pathlib.Path(dirpath) / filename for filename in filenames
-        )
-        shas += (hashlib.sha1(file.read_bytes()).digest() for file in files)
-    return hashlib.sha1(b"".join(sorted(shas))).digest()
 
 
 def expected_num_members_group_assertion(expected_num_members):
