+++ system_tests/gitlab/test_gitlab_system.py
@@ -7,6 +7,13 @@ import pytest
 import repobee_plug as plug
 import repobee_testhelpers
 
+from repobee_testhelpers.funcs import hash_directory
+from repobee_testhelpers._internal.docker import (
+    run_in_docker_with_coverage,
+    run_in_docker,
+    VOLUME_DST,
+)
+
 import _repobee.ext
 import _repobee.command.peer
 import _repobee.ext.gitlab
@@ -23,7 +30,6 @@ from _helpers.asserts import (
     assert_cloned_repos,
 )
 from _helpers.const import (
-    VOLUME_DST,
     BASE_DOMAIN,
     LOCAL_DOMAIN,
     ORG_NAME,
@@ -41,10 +47,7 @@ from _helpers.const import (
 )
 from _helpers.helpers import (
     api_instance,
-    run_in_docker_with_coverage,
-    run_in_docker,
     update_repo,
-    hash_directory,
     expected_num_members_group_assertion,
 )
 
@@ -73,7 +76,8 @@ class TestClone:
         assert_cloned_repos(STUDENT_TEAMS, assignment_names, tmpdir)
 
     def test_clone_twice(self, with_student_repos, tmpdir, extra_args):
-        """Cloning twice in a row should have the same effect as cloning once.
+        """Cloning twice in a row should have the same effect as cloning
+        once.
         """
         command = " ".join(
             [
